Select: 가독성 향상을 위한 리펙토링
