+++ src/components/select/select.tsx
@@ -1,38 +1,7 @@
 import React, { useEffect, useState, useRef } from 'react';
-import type { Value, Option, CustomSelectProps, SelectProps } from '../../types/select.interface';
-import styled from '@emotion/styled';
-import useElementHeight from './hooks/useElementHeight';
+import type { Value, SelectProps } from '../../types/select.interface';
+import CustomSelect from './custom-select';
 import SelectUseOption from './select-use-option';
-
-const CustomSelect = styled.div<CustomSelectProps>`
-  position: relative;
-  & > .options {
-    opacity: ${({ isOptionVisible }) => (isOptionVisible ? '1' : '0')};
-    z-index: ${({ isOptionVisible }) => (isOptionVisible ? '1' : '-1')};
-
-    position: absolute;
-    ${({ customOptionHeight, size }) => {
-      if (size) {
-        return `
-          overflow: scroll;
-          height: ${customOptionHeight * size}px;
-        `;
-      }
-      return '';
-    }}
-  }
-
-  & > .wrapper {
-    display: ${({ isOptionVisible }) => (isOptionVisible ? 'block' : 'none')};
-    position: fixed;
-    z-index: 0;
-    width: 100vw;
-    height: 100vh;
-    position: fixed;
-    top: 0;
-    left: 0;
-  }
-`;
 
 const Select = ({
   options,
@@ -44,8 +13,6 @@ const Select = ({
   setIsAllSelect,
 }: SelectProps) => {
   const [selectedValue, setSelectedValue] = useState<Value[]>(multiple ? [] : [options[0].value]);
-  const [isOptionVisible, setIsOptionVisible] = useState(false);
-  const [customOptionHeight, customOptionRef] = useElementHeight();
   const isAllSelected = options.length === selectedValue.length;
 
   useEffect(() => {
@@ -63,46 +30,10 @@ const Select = ({
     }
   }, [isAllSelect]);
 
-  const previewClickHandler = () => {
-    setIsOptionVisible((state) => {
-      return !state;
-    });
-  };
-
-  const setValueHandler = (e: React.ChangeEvent<HTMLSelectElement>) => {
-    if (multiple) {
-      const options = e.target.options;
-      const selectedValues = [...options]
-        .filter((option) => option.selected)
-        .map((option) => {
-          return option.value;
-        });
-
-      setSelectedValue(selectedValues);
-      return;
-    }
-    setSelectedValue([e.target.value]);
-  };
-
-  const optionClickHandler = (option: Option) => () => {
-    if (multiple) {
-      setSelectedValue((state) => {
-        const optionIdx = state.indexOf(option.value);
-        if (optionIdx !== -1) {
-          return [...state.slice(0, optionIdx), ...state.slice(optionIdx + 1, state.length)];
-        }
-        return [...state, option.value];
-      });
-      return;
-    }
-    setSelectedValue([option.value]);
-    setIsOptionVisible(false);
-  };
-
   if (isOptionUse) {
     return (
       <SelectUseOption
-        setValueHandler={setValueHandler}
+        setSelectedValue={setSelectedValue}
         multiple={multiple}
         options={options}
         isAllSelect={isAllSelect}
@@ -110,34 +41,14 @@ const Select = ({
       />
     );
   }
-
   return (
-    <div className="fc-select">
-      <div className="preview" onClick={previewClickHandler}>
-        {selectedValue}
-      </div>
-
-      <CustomSelect customOptionHeight={customOptionHeight} size={size} isOptionVisible={isOptionVisible}>
-        <div className="options">
-          {options.map((option, idx) => (
-            <div
-              key={option.value}
-              onClick={optionClickHandler(option)}
-              className={selectedValue.includes(option.value) ? 'selected' : ''}
-              ref={idx === 0 ? customOptionRef : null}
-            >
-              {option.innerHTML}
-            </div>
-          ))}
-        </div>
-        <div
-          className="wrapper"
-          onClick={() => {
-            setIsOptionVisible(false);
-          }}
-        ></div>
-      </CustomSelect>
-    </div>
+    <CustomSelect
+      selectedValue={selectedValue}
+      setSelectedValue={setSelectedValue}
+      options={options}
+      size={size}
+      multiple={multiple}
+    />
   );
 };
